Python/98/requirements.txt
Flask
loguru
python-telegram-bot
Requests
selenium
webdriver_manager
